requirements.txt
pika==0.10.0
ckanext-harvest=0.2
simplejson==3.3.1
pymongo==2.8.1
configparser==3.5.0b2
requests==2.7.0
python-dateutil==1.5
hashlib==20081119
termcolor==1.1.0
beautifulsoup4==4.3.2
pytz==2014.7
goslate==1.3.0
planar==0.4
LEPL==5.1.3
urllib3==1.12
Whoosh==2.7.0
jellyfish==0.5.1
